Add API with collection endpoint
Adds the API struct and a single endpoint object for collections. The
collection struct exposes 3 functions, one for getting all owned games
for a certain user, one for wishlisted games, and finally one which will
accept a querybuilder, so the user can write a custom query.

Mostly this is all for proof of concept. The querybuilder needs fleshing
out with all the parameters. The model itself also needs all fields
added and documentation comments are needed for everything, as well as
tests.