Return API error over parse error if possible
The XML api will sometimes return a 200 in the case of an error, with an
error message wrapped in XML. For example when the username does not
exist. In this case, rather than just return the serde_xml_rs parse
error. We want to try to deserialise the response again as an API error.
If this is possible we return a more appropriate API error with the
message given to us from the server.