Change model derived traits
All models now derive Clone, Debug, and Deserialize. But not
Serialize and nor PartialEq. Serialize is excluded because there is no
reason to reserialize them into xml, and I don't think there would be
a guarantee that it would be the same format that it started as.

As for PartialEq, I haven't yet encountered a need for it, and I would
like to keep things slim and simple as possible until I needed.